Fixes and debugging

+++ install.sh
@@ -1,13 +1,14 @@
-read -p "enter the disk on which to apply(e.g. /dev/sdx): " disk
-read -p "enter the efi partition size(e.g. 512 Integer):" efi_size
-read -p "enter the swap partition size(e.g. 4096 Integer)" swap_size
-read -p "enter the hostname for the setup(e.g. anyname): " hostname
-read -p "enter the password for root(e.g. @p455w0rd123): " password
-read -p "enter the new user's name: " username
-read -p "password for the new user: " userpass
+read -p "enter the disk on which to apply(e.g. /dev/sdx) : " disk
+read -p "enter the efi partition size(e.g. 512 Integer) :" efi_size
+read -p "enter the swap partition size(e.g. 4096 Integer) :" swap_size
+read -p "enter the hostname for the setup(e.g. anyname) : " hostname
+read -p "enter the password for root(e.g. @p455w0rd123) : " password
+read -p "enter the new user's name : " username
+read -p "password for the new user : " userpass
 
 umount -R /mnt || true
 umount -R "${disk}"* 2>/dev/null || true
+umount -R "${disk}" 2>/dev/null || true
 
 if ! [[ -b "$disk" ]]; then
   echo "Specified invalid disk. Please specify correct disk.";
@@ -27,18 +28,26 @@ fi
 disk_size=$(echo "$(lsblk -bndo SIZE ${disk})/1024/1024" | bc) # MB
 EFI_PART=$efi_size # MB
 SWAP_PART=$swap_size # MB
+echo "[ DISK SIZE\t\t<==>\t\t[$disk_size] ] \n[ EFI PARTITION\t\t<==>\t\t[$EFI_PART] ] \n[ SWAP PARTITION\t\t<==>\t\t[$SWAP_PART] ]";
 
-
-umount ${disk} 2>/dev/null
 parted ${disk} mklabel gpt # clearing partition
+if [[ $? -ne 0 ]]; then echo "Failed while clearing partition"; exit 1; fi
+
+
 parted ${disk} mkpart primary fat32 1MiB ${EFI_PART}MiB # creating EFI
+if [[ $? -ne 0 ]]; then echo "Failed while creating EFI"; exit 1; fi
+
 parted ${disk} set 1 esp on # assigning the EFI flag here
+if [[ $? -ne 0 ]]; then echo "Failed while assigning the EFI flag"; exit 1; fi
 
 parted ${disk} mkpart primary linux-swap ${EFI_PART}MiB $(echo ${EFI_PART}+${SWAP_PART} | bc)MiB # creating the swap partition here
+if [[ $? -ne 0 ]]; then echo "Failed while creating the swap partition"; exit 1; fi
 
 parted ${disk} mkpart primary ext4 $(echo "${EFI_PART}+${SWAP_PART}" | bc)MiB ${disk_size}MiB # creating root partition
+if [[ $? -ne 0 ]]; then echo "Failed while creating the ROOT partition"; exit 1; fi
 
 partprobe ${disk} # inform the OS about changed partition table
+if [[ $? -ne 0 ]]; then echo "Failed while applying changes for the partitions"; exit 1; fi
 
 #formatting partitions
 mkfs.fat -F32 ${disk}1
@@ -54,7 +63,7 @@ pacstrap -K /mnt base linux linux-firmware hyprland waybar wofi swaybg swaylock 
 
 # creating the table and saving inside the /mnt/etc/fstab
 genfstab -U /mnt >> /mnt/etc/fstab
-
+if [[ $? -ne 0 ]]; then echo "Failed while saving the generated file system table for permanent changes"; exit 1; fi
 
 arch-chroot /mnt << EOF
 pacman -Sy
